+++ CHANGELOG.md
@@ -34,6 +34,9 @@
 - Bugfix: A potential deadlock situation is fixed that sometimes caused the user daemon to hang when the user
   was logged in. 
 
+- Feature: The scout reports will now include additional metadata coming from environment variables starting with
+  `TELEPRESENCE_REPORT_`.
+
 ### 2.4.0 (August 4, 2021)
 
 - Feature: There is now a native Windows client for Telepresence.